+++ inst/superbShiny/App.R
@@ -14,7 +14,7 @@ library(foreign) # for read.spss
 library(stringr) # for str_remove_all and str_replace_all
 library(dplyr)
 
-appversion <- "App version 3.5; shipped with superb 0.95.13"
+appversion <- "App version 3.6; shipped with superb 0.95.16"
 
 ##########################################################
 ##########################################################
@@ -344,7 +344,7 @@ thePage <- fluidPage(
                     selectInput("superbPlotorder", "Select the order in which factors are plotted", 
                         multiple = TRUE, choices=c(""), selected="" ),
                     selectInput("superbLayout", "Select the plots' layout",
-                        choices=c("bar","point","line","lineBand","pointjitter","pointjitterviolin","pointindividualline","raincloud","halfwidthline","lineBand"),
+                        choices=c("bar","point","line","pointjitter","pointjitterviolin","pointindividualline","raincloud","halfwidthline","boxplot","lineBand","corset"),
                         multiple = FALSE, selectize = FALSE),
                     strong("Specific graphic attributes (comma separated)"),br(),
                     div(style="display:inline-block", # if one choice is named, they must all be?