+++ tests/testthat/test_superbPlot.R
@@ -842,3 +842,56 @@ test_that("Tryon vs. difference", {
 
 
 
+#########################################
+# testing multiple formats!
+#########################################
+
+test_that("Testing corset plot", {
+	old <- options() 
+	on.exit(options(old)) 
+    options("superb.feedback" = c('none'))
+
+    ## corset plot
+    dta <- GRD(SubjectsPerGroup = 50, WSFactors = "moment(2)", Effects = list("moment"=slope(3)))
+    plt <- superbPlot(dta, WSFactors = "moment(2)", variables = c("DV.1","DV.2"),
+      plotStyle = "corset" )
+    expect_equal( "ggplot" %in% class(plt), TRUE)
+
+    plt <- superbPlot(dta, WSFactors = "moment(2)", variables = c("DV.1","DV.2"),
+        plotStyle    = "corset", 
+        lineParams   = list(colorize=TRUE),
+        violinParams = list(fill = "green", alpha = 0.2 ) 
+    ) + theme_bw() + 
+    theme(axis.line.y = element_line(color="black"), legend.position=c(0.1,0.75), panel.border = element_blank() ) +
+    scale_color_manual('Direction\n of change', values=c("red","gray50"), labels=c('decreasing', 'increasing'))
+    expect_equal( "ggplot" %in% class(plt), TRUE)
+
+    # restores default information
+    options("superb.feedback" = c('design','warnings','summary'))
+})
+
+
+test_that("Testing local decorrelation plot", {
+	old <- options() 
+	on.exit(options(old)) 
+    options("superb.feedback" = c('none'))
+
+    #test
+    X <- GRD(SubjectsPerGroup = 50, WSFactors = "time(10)", Effects = list("time"=extent(3)))
+    go <- function(m) {
+        superbPlot(X, 
+            WSFactor = "time(10)",
+            variables = names(X)[-1],
+            adjustments = list(decorrelation = m),
+            plotStyle = "lineBand"
+        )+ylim(c(-6,6))
+    }
+    plt <- go("LD2")
+    expect_equal( "ggplot" %in% class(plt), TRUE)
+
+    # restores default information
+    options("superb.feedback" = c('design','warnings','summary'))
+})
+
+
+